+++ demos/python/sdk_wireless_camera_control/open_gopro/api/v1_0/wifi_commands.py
@@ -6,11 +6,11 @@
 from __future__ import annotations
 import logging
 from pathlib import Path
-from typing import Any, Optional, Type, Dict, Callable
+from typing import Any, Optional, Type, Dict
 
 from open_gopro.communication_client import GoProWifi
 from open_gopro.constants import SettingId, StatusId
-from open_gopro.responses import BytesParserBuilder, GoProResp, JsonParser
+from open_gopro.responses import BytesParserBuilder, JsonParser
 from open_gopro.api.builders import WifiSetting, WifiGetJsonNoParams, WifiGetJsonWithParams, WifiGetBinary
 from .params import ParamsV1_0 as Params
 
@@ -26,7 +26,7 @@ class WifiCommandsV1_0:
     All of these return a GoProResp
 
     Args:
-        communicator (GoProWifi): [description]
+        communicator (GoProWifi):  Adapter to read / write commands
     """
 
     # TODO refactor this to reuse response parsing
@@ -53,8 +53,7 @@ class WifiCommandsV1_0:
                     identifier = id_map(int(k))
                     try:
                         parser = additional_parsers[identifier]
-                        # GreedyBytes, GreedyString, etc can't be built since they don't have a length
-                        val = v if "Greedy" in str(parser) else parser.parse(parser.build(v))
+                        val = parser.parse(parser.build(v))
                     except KeyError:
                         logger.warning(f"unparsed {name}: {identifier}")
                         val = v
@@ -69,41 +68,39 @@ class WifiCommandsV1_0:
     def __init__(self, communicator: GoProWifi):
         self.communicator = communicator
 
-        # ======================== Classes used to build commands
-
-        class GetJsonFromInt(WifiGetJsonWithParams[int]):
-            ...
-
-        class GetJsonFromStr(WifiGetJsonWithParams[str]):
-            ...
-
-        class GetJsonFromToggle(WifiGetJsonWithParams[Params.Toggle]):
-            ...
-
-        class GetJsonFromPreset(WifiGetJsonWithParams[Params.Preset]):
-            ...
-
-        class GetJsonFromPresetGroup(WifiGetJsonWithParams[Params.PresetGroup]):
-            ...
-
         class CameraFileToLocalFile(WifiGetBinary):
             def __call__(self, /, camera_file: str, local_file: Optional[Path] = None) -> Path:
                 return super().__call__(camera_file=camera_file, local_file=local_file or camera_file)
 
         # ======================================== Commands
 
-        self.set_digital_zoom: Callable[[int], GoProResp] = GetJsonFromInt(
+        self.set_digital_zoom = WifiGetJsonWithParams[int](
             communicator, "gopro/camera/digital_zoom?percent={}"
         )
-        """Set digital zoom in percent (0 to 100)."""
+        """Set digital zoom in percent.
+
+        Args:
+            value (int): Desired zoom as a percentage
+
+        Returns:
+            GoProResp: command status
+        """
 
         self.get_camera_state = WifiGetJsonNoParams(
             communicator, "gopro/camera/state", response_parser=WifiCommandsV1_0._ParseCameraState()
         )
-        """Get camera status and settings."""
+        """Get camera status and settings.
+
+        Returns:
+            GoProResp: command status
+        """
 
         self.set_keep_alive = WifiGetJsonNoParams(communicator, "/gopro/camera/keep_alive")
-        """Send the keep alive signal to maintain the connection."""
+        """Send the keep alive signal to maintain the connection.
+
+        Returns:
+            GoProResp: command status
+        """
 
         self.get_gpmf_data = CameraFileToLocalFile(communicator, "gopro/media/gpmf?path=100GOPRO/{}")
         """Get GPMF data for a file.
@@ -118,11 +115,22 @@ class WifiCommandsV1_0:
             Path: Path to local_file that output was written to
         """
 
-        self.get_media_info = GetJsonFromStr(communicator, "gopro/media/info?path=100GOPRO/{}")
-        """Get media info for a file."""
+        self.get_media_info = WifiGetJsonWithParams[str](communicator, "gopro/media/info?path=100GOPRO/{}")
+        """Get media info for a file.
+
+        Args:
+            value (str): Media file to get info for
+
+        Returns:
+            GoProResp: command status and media info as JSON
+        """
 
         self.get_media_list = WifiGetJsonNoParams(communicator, "gopro/media/list")
-        """Get a list of media on the camera."""
+        """Get a list of media on the camera.
+
+        Returns:
+            GoProResp: command status and media list as JSON
+        """
 
         self.get_screennail = CameraFileToLocalFile(communicator, "gopro/media/screennail?path=100GOPRO/{}")
         """Get screennail for a file.
@@ -150,28 +158,75 @@ class WifiCommandsV1_0:
             Path: Path to local_file that output was written to
         """
 
-        self.set_turbo_mode = GetJsonFromToggle(communicator, "/gopro/media/turbo_transfer?p={}")
-        """Enable or disable Turbo transfer mode."""
+        self.set_turbo_mode = WifiGetJsonWithParams[Params.Toggle](
+            communicator, "/gopro/media/turbo_transfer?p={}"
+        )
+        """Enable or disable Turbo transfer mode.
+
+        Args:
+            value (Params.Toggle): enable / disable turbo mode
+
+        Returns:
+            GoProResp: command status
+        """
 
         self.get_open_gopro_api_version = WifiGetJsonNoParams(communicator, "gopro/version")
-        """Get Open GoPro version."""
+        """Get Open GoPro version.
+
+        Returns:
+            GoProResp: command status and version separated into major and minor fields
+        """
 
         self.get_preset_status = WifiGetJsonNoParams(communicator, "gopro/camera/presets/get")
-        """Get current Preset status."""
+        """Get current Preset status.
 
-        self.set_preset = GetJsonFromPreset(communicator, "gopro/camera/presets/load?id={}")
+        Returns:
+            GoProResp: command status and current preset status as JSON
+        """
+
+        self.set_preset = WifiGetJsonWithParams[Params.Preset](communicator, "gopro/camera/presets/load?id={}")
         """Set the active preset."""
 
-        self.set_preset_group = GetJsonFromPresetGroup(communicator, "gopro/camera/presets/set_group?id={}")
-        """Set the active preset group."""
+        self.set_preset_group = WifiGetJsonWithParams[Params.PresetGroup](
+            communicator, "gopro/camera/presets/set_group?id={}"
+        )
+        """Set the active preset group.
+
+        The most recently used Preset in this group will be set.
+
+        Args:
+            value (Params.PresetGroup): desired Preset Group
+
+        Returns:
+            GoProResp: command status
+        """
 
         self.start_preview_stream = WifiGetJsonNoParams(communicator, "gopro/camera/stream/start")
-        """Start the preview stream."""
+        """Start the preview stream.
+
+        Returns:
+            GoProResp: command status
+        """
 
         self.stop_preview_stream = WifiGetJsonNoParams(communicator, "gopro/camera/stream/stop")
-        """Stop the preview stream."""
+        """Stop the preview stream.
+
+        Returns:
+            GoProResp: command status
+        """
 
         self.get_telemetry = CameraFileToLocalFile(communicator, "gopro/media/telemetry?path=100GOPRO/{}")
+        """Download the telemetry data for a camera file and store in a local file.
+
+        If local_file is none, the output location will be the same name as the camera_file.
+
+        Args:
+            camera_file (str): filename on camera to operate on
+            local_file (Optional[Path], optional): Location on computer to write output. Defaults to None.
+
+        Returns:
+            Path: Path to local_file that output was written to
+        """
 
         self.download_file = CameraFileToLocalFile(communicator, "videos/DCIM/100GOPRO/{}")
         """Download a video from the camera to a local file.
@@ -197,6 +252,8 @@ class WifiSettingsV1_0:
 
     Args:
         communicator (GoProWifi): Adapter to read / write settings
+        params: (Type[Params]): the set of parameters to use to build the settings
+        endpoint (str): the endpoint to write to to set a setting
     """
 
     class Iterator:
@@ -234,46 +291,27 @@ class WifiSettingsV1_0:
         self.endpoint = endpoint
         self.params = params  # Not currently used but will be in the future
 
-        class Resolution(WifiSetting[Params.Resolution]):
-            ...
-
-        self.resolution = Resolution(communicator, SettingId.RESOLUTION)
+        self.resolution = WifiSetting[Params.Resolution](communicator, SettingId.RESOLUTION)
         """Resolution. Set with :py:class:`open_gopro.Params.Resolution`"""
 
-        class FPS(WifiSetting[Params.FPS]):
-            ...
-
-        self.fps = FPS(communicator, SettingId.FPS)
+        self.fps = WifiSetting[Params.FPS](communicator, SettingId.FPS)
         """Frames per second. Set with :py:class:`open_gopro.Params.FPS`"""
 
-        class AutoOffTime(WifiSetting[Params.AutoOff]):
-            ...
+        self.auto_off = WifiSetting[Params.AutoOff](communicator, SettingId.AUTO_OFF)
+        """Set the auto off time. Set with :py:class:`Params.AutoOff`"""
 
-        self.auto_off = AutoOffTime(communicator, SettingId.AUTO_OFF)
-        """Set the auto off time."""
-
-        class VideoFOV(WifiSetting[Params.VideoFOV]):
-            ...
-
-        self.video_field_of_view = VideoFOV(communicator, SettingId.VIDEO_FOV)
+        self.video_field_of_view = WifiSetting[Params.VideoFOV](communicator, SettingId.VIDEO_FOV)
         """Video FOV. Set with :py:class:`open_gopro.Params.FieldOfView`"""
 
-        class PhotoFOV(WifiSetting[Params.PhotoFOV]):
-            ...
-
-        self.photo_field_of_view = PhotoFOV(communicator, SettingId.PHOTO_FOV)
+        self.photo_field_of_view = WifiSetting[Params.PhotoFOV](communicator, SettingId.PHOTO_FOV)
         """Photo FOV. Set with :py:class:`open_gopro.Params.FieldOfView`"""
 
-        class MultishotFOV(WifiSetting[Params.MultishotFOV]):
-            ...
-
-        self.multi_shot_field_of_view = MultishotFOV(communicator, SettingId.MULTI_SHOT_FOV)
+        self.multi_shot_field_of_view = WifiSetting[Params.MultishotFOV](
+            communicator, SettingId.MULTI_SHOT_FOV
+        )
         """Multi-shot FOV. Set with :py:class:`open_gopro.Params.FieldOfView`"""
 
-        class MaxLensMode(WifiSetting[Params.MaxLensMode]):
-            ...
-
-        self.max_lens_mode = MaxLensMode(communicator, SettingId.MAX_LENS_MOD)
+        self.max_lens_mode = WifiSetting[Params.MaxLensMode](communicator, SettingId.MAX_LENS_MOD)
         """Enable / disable max lens mod. Set with :py:class:`open_gopro.Params.MaxLensMode`"""
 
     def __iter__(self) -> Iterator: