+++ demos/python/sdk_wireless_camera_control/docs/changelog.rst
@@ -9,6 +9,13 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+0.7.1 (December-16-2021)
+-------------------------
+* Add global behaviors commands and camera control status
+* Add register / unregister all for settings and statuses
+* Add max lens setting
+* Improve API documentation
+
 0.7.0 (October-27-2021)
 -------------------------
 * Add video performance mode functionality